+++ sites/nau/src/backend/nau/settings.py
@@ -599,7 +599,13 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     SESSION_ENGINE = values.Value("django.contrib.sessions.backends.cache")
 
     # Sentry
-    SENTRY_DSN = values.Value(None, environ_name="SENTRY_DSN")
+    SENTRY_DSN = values.Value(None, environ_name="SENTRY_DSN", environ_prefix=None)
+    SENTRY_ENVIRONMENT = values.Value(
+        __name__.lower(), environ_name="SENTRY_ENVIRONMENT", environ_prefix=None
+    )
+    SENTRY_ENABLE_TRACKING = values.Value(
+        False, environ_name="SENTRY_ENABLE_TRACKING", environ_prefix=None
+    )
 
     # Admin
     # - Django CMS
@@ -684,9 +690,10 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         if cls.SENTRY_DSN is not None:
             sentry_sdk.init(
                 dsn=cls.SENTRY_DSN,
-                environment=cls.__name__.lower(),
+                environment=cls.SENTRY_ENVIRONMENT or cls.__name__.lower(),
                 release=get_release(),
                 integrations=[DjangoIntegration()],
+                enable_tracing=cls.SENTRY_ENABLE_TRACKING,
             )
             with sentry_sdk.configure_scope() as scope:
                 scope.set_extra("application", "backend")